requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
alembic
dotenv
pydantic
pydantic-settings
bcrypt<4.0
